Add additional tests (#13)
* Move the outputs to the right place
* Add more tests
* Add additional tests
* Remove the implicit ctx and key from the input

This commit removes the implicit `ctx` and `key` that are
passed through to the machine's input on create.
While in principle it is a nice idea it causes the ctx (restate context)
to be serialized and persisted into the machine's snapshot, for example
in the following snippet:
```
 context: ({ input }) => ({
    counts: input, // <-- implicitly restate ctx
  }),
```

This ends up with a serialization error at runtime.
The restate key however is accessible by using `self.id`

* Replace key -> self.id in the example